+++ src/test/resources/testResult/featureCollection/monitoring-locations/counties_ak_05051.json
