+++ src/test/resources/testResult/featureCollection/monitoring-locations/monitoring_location_number_USGS-343204093005501.json
@@ -41,19 +41,19 @@
       "links": [
         {
           "rel": "self",
-          "href": "http://localhost:8080/api/observations/collections/monitoring-locations/items/USGS-343204093005501?f=json",
+          "href": "https://unittest/api/observations/collections/monitoring-locations/items/USGS-343204093005501?f=json",
           "type": "application/geo+json",
           "title": "This document as GeoJSON"
         },
         {
           "rel": "collection",
-          "href": "http://localhost:8080/api/observations/collections/monitoring-locations?f=json",
+          "href": "https://unittest/api/observations/collections/monitoring-locations?f=json",
           "type": "application/json",
           "title": "NWIS Monitoring Locations"
         },
         {
           "rel": "collection",
-          "href": "http://localhost:8080/api/observations/collections/AHS?f=json",
+          "href": "https://unittest/api/observations/collections/AHS?f=json",
           "type": "application/json",
           "title": "Arkansas Hot Springs National Park Network"
         }
@@ -71,13 +71,13 @@
       "type": "application/geo+json",
       "rel": "self",
       "title": "This document as GeoJSON",
-      "href": "http://localhost:8080/api/observations/collections/monitoring-locations/items?f=json&startIndex=0&limit=10000"
+      "href": "https://unittest/api/observations/collections/monitoring-locations/items?f=json&startIndex=0&limit=10000"
     },
     {
       "type": "application/json",
       "rel": "collection",
       "title": "NWIS Monitoring Locations",
-      "href": "http://localhost:8080/api/observations/collections/monitoring-locations?f=json"
+      "href": "https://unittest/api/observations/collections/monitoring-locations?f=json"
     }
   ]
 }
