+++ src/test/resources/testResult/featureCollection/monitoring-locations/states_mi.json
@@ -41,13 +41,13 @@
       "links": [
         {
           "rel": "self",
-          "href": "http://localhost:8080/api/observations/collections/monitoring-locations/items/USGS-04027940?f=json",
+          "href": "https://unittest/api/observations/collections/monitoring-locations/items/USGS-04027940?f=json",
           "type": "application/geo+json",
           "title": "This document as GeoJSON"
         },
         {
           "rel": "collection",
-          "href": "http://localhost:8080/api/observations/collections/monitoring-locations?f=json",
+          "href": "https://unittest/api/observations/collections/monitoring-locations?f=json",
           "type": "application/json",
           "title": "NWIS Monitoring Locations"
         }
@@ -65,13 +65,13 @@
       "type": "application/geo+json",
       "rel": "self",
       "title": "This document as GeoJSON",
-      "href": "http://localhost:8080/api/observations/collections/monitoring-locations/items?f=json&startIndex=0&limit=10000"
+      "href": "https://unittest/api/observations/collections/monitoring-locations/items?f=json&startIndex=0&limit=10000"
     },
     {
       "type": "application/json",
       "rel": "collection",
       "title": "NWIS Monitoring Locations",
-      "href": "http://localhost:8080/api/observations/collections/monitoring-locations?f=json"
+      "href": "https://unittest/api/observations/collections/monitoring-locations?f=json"
     }
   ]
 }
